Move mods {a,}sync -> {a,}sync_client
